+++ src/game/Player.java
@@ -14,7 +14,7 @@ public class Player extends Character {
     
     public Player(String name, Race race, Proffesion proffesion, 
             Inventory inventory, Atributes atributes, int healthPoints, 
-            int level, byte alignment, int exp) {
+            int level, int alignment, int exp) {
         super(name, race, proffesion, inventory, atributes, 
                 healthPoints, level, alignment);
         this.exp = exp;
